+++ source/renderer/zettlr-editor.js
@@ -61,6 +61,7 @@ require('./assets/codemirror/zettlr-plugin-render-images.js')
 require('./assets/codemirror/zettlr-plugin-render-links.js')
 require('./assets/codemirror/zettlr-plugin-render-tasks.js')
 require('./assets/codemirror/zettlr-plugin-render-iframes.js')
+require('./assets/codemirror/zettlr-plugin-render-math.js')
 require('./assets/codemirror/zettlr-plugin-markdown-header-classes.js')
 
 // Finally CodeMirror itself
@@ -158,6 +159,7 @@ class ZettlrEditor {
       // cursor changes its position as well then) or when the cursor moves.
       this._cm.execCommand('markdownRenderImages') // Render images
       this._cm.execCommand('markdownRenderIframes') // Render iFrames
+      this._cm.execCommand('markdownRenderMath') // Render equations
       this._cm.execCommand('markdownRenderLinks') // Render links
       this._cm.execCommand('markdownRenderTasks') // Render tasks
       this._cm.execCommand('markdownHeaderClasses') // Apply heading line classes